requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
pyjwt[crypto]==2.8.0
aio-pika==9.3.1
python-multipart==0.0.6
httpx==0.26.0
python-dotenv
psycopg2-binary
pytz
passlib
python-multipart
bcrypt==4.0.1
python-jose[cryptography]